+++ build.mjs
@@ -1,10 +1,7 @@
-import pkg from './package.json' assert { type: 'json' };
-import { builtinModules } from 'module';
 import { build } from 'esbuild';
 
 /** @type {import('esbuild').BuildOptions} */
 const options = {
-  external: [...Object.keys(pkg.devDependencies), ...builtinModules],
   entryPoints: { index: './src/index.ts' },
   sourcemap: false,
   logLevel: 'debug',